+++ include/display/Sprite.hpp
@@ -34,14 +34,14 @@ class Sprite : public Image {
 		
 		void addAnimation(std::initializer_list<u16> frames, u16 delay);
 		
-		void drawFrame(s16 x, s16 y, u16 frame);
+		void drawFrame(float x, float y, u16 frame);
 		
 		void resetAnimation(u16 anim, u16 frame = 0);
 		void startAnimation(u16 anim);
 		void stopAnimation(u16 anim);
 		u16 animationCurrentFrame(u16 anim);
 		bool animationAtEnd(u16 anim);
-		void playAnimation(s16 x, s16 y, u16 anim);
+		void playAnimation(float x, float y, u16 anim);
 		
 		static bool pause;
